+++ src/database/in_memory/mod.rs
@@ -1,0 +1,1 @@
+pub mod in_memory_db;
