+++ src/database/mod.rs
@@ -1,10 +1,11 @@
+pub mod db;
+pub mod in_memory;
 pub mod options;
+pub mod reader_writer;
 pub mod traits;
 pub mod types;
 pub mod utils;
 
 mod db_base;
-mod reader_base;
 
 pub use db_base::DB;
-pub use reader_base::ReaderBase;
