+++ src/sparse_merkle_tree/mod.rs
@@ -1,0 +1,5 @@
+pub mod in_memory_smt;
+pub mod smt;
+pub mod smt_db;
+
+pub use smt::{Proof, QueryProof, SparseMerkleTree, UpdateData};
